refactor: Lotto 클래스 삭제
