+++ src/domain/constants/numbers.js
@@ -1,0 +1,7 @@
+const NUMBERS = {
+  lottoNumbersLength: 6,
+  minLottoNumber: 1,
+  maxLottoNumber: 45,
+};
+
+export default NUMBERS;
